r75: - Reduce usage of queue_lock to the minimum necessary for the block layer, simplifying the locking regime.  Most functions run without the queue lock and take it when they need it.
- Give the chunkdata table its own lock, so those functions no longer need
to be synchronized by the caller

- Give dev->setup_sg its own lock

- Reuse io structures for multiple-chunk requests.  Get rid of the
pending_requests queue, since it's no longer possible for a multiple-chunk
request to run out of memory partway through.  Get rid of lowmem_recover().

- Split convergent_handle_request() into convergent_setup_io() and
convergent_process_io().  All potentially failing operations go in the
first function, and the second prepares io for submission (and may
be called multiple times for a multiple-chunk request).  This eliminates
the need to distinguish requests that have reservations from those that
don't - all requests with an io have reservations and all the memory they'll
ever need, and the rest get stuck back on the queue.

- Move -ENOMEM handling back into convergent_request(), since that's the
only caller of convergent_setup_io()

- Move io_cleaner() later in convergent.c

- Move code to free a convergent_io to its own function

- Rename convergent_callback() and convergent_endio() for clarity