r78: - Convert explicit hex defines in convergent.h to enums and bitshifts
- Remove support for orig_req mapped sg's larger than MAX_SEGS_PER_IO,
to simplify the code as it's being rearranged; we'll need to add this back
in later

- Allow chunks from multiple-chunk ios to complete in parallel.  Have
each convergent_io contain an array of convergent_io_chunk, each of
which carries all the chunk-specific state.  Handle out-of-order completion.
With this change, we no longer iterate a request through
convergent_process_io(); each io is set up exactly once.  Also we get
better performance, since we're not defeating the chunk store's I/O
elevator.