r92: - Remove large per-device scatterlist.
Rationale: Support for a segment count > MAX_SEGS_PER_IO is broken, and
either needs to be fixed or removed.  It'd be possible to re-add support;
we'd have to treat the io as a window into part of a request, and when we
complete everything in the window, update the io_chunk state to point to the
new window.  But, since we now share chunk pages between ios, multiple ios
to a single chunk doesn't have the performance penalty it used to.  The only
remaining performance penalty is the fact that if an entire chunk is being
written, but over multiple requests, we do a read-in that we don't need to
do.  (If we start tracking sector ranges touched by each io so that we can
do emergency writeback without reading in a chunk first, we might be able to
work around this issue by walking the pending queue and reading that data.)
So, the performance gain doesn't seem worth the extra complexity, and I'm
removing the feature for now.