r93: - Select cipher, hash, compression algorithms based on parameters passed from userspace rather than hardcoding the algo selection
- Retrieve the hash length from the algo module rather than hardcoding
a constant.  Pass the length back to userspace during registration.

- Redo usermsg functions in chunkdata.c to provide a more extensible
interface to chardev.c

- Dequeue chunks for userspace with a two-step process to allow for the
fact that we might need to requeue the chunk if the copy_to_user() fails

- Replace the proliferation of chunkdata flags with an explicit state
machine.  The downside is that we're now forced to work in a more-or-less
linear model.  The upside is that we're now forced to work in a linear
model.  The code is much easier to understand this way.

- Replace ISR_MSG flags with ISR_MSGTYPE message types, for similar reasons.
We really don't need multiple operations per message.