r112: - Add refcount of items that require userspace to be around, and user_get() and user_put() functions to manipulate it.  These items are: each open of the block device, each busy chunk, and each pending chunk_io.  Do not allow the unregister ioctl to succeed unless the refcount is zero. Allow userspace to check for a zero refcount by giving the POLLPRI flag to poll().  Wake up waiting processes when the refcount reaches zero.
This means we now effectively have two refcounts.  user_get()/put() is the
weaker one; when it reaches zero, userspace is allowed to go away (in cases
where it asks nicely first).  convergent_dev_get()/put() is stronger: when it
reaches zero, the convergent_dev is deallocated.  The difference lies in two
facts: userspace can go away when the user count is non-zero if it really
wants to (through close() or by crashing), and there are some references
held by the system (e.g., to sysfs files) which are not included in the user
count.

- Get rid of separate busy-chunk and pending-chunk-io counters.  This
means that chunkdata's reference to convergent_dev will not be released
until all fds to the block device are closed, in addition to the other
requirements.