r173: - Limit cachesize to 10% of system RAM, rather than a static number of buffers
- Prevent multiple convergent_devs from collectively allocating more than
30% of system RAM, to prevent a DoS in which users with access to
/dev/openisrctl could fill memory with (unswappable) chunk buffers

- Allow chunk buffers to be in high memory.  Most drivers for the underlying
chunk device will be using 32-bit DMA, so the device will be able to read
the buffers even though they may not be in kernel-mapped memory -- so in
most cases this won't result in the creation of bounce buffers.  Bounce
buffers may still be needed on 64-bit or large 32-bit systems with 32-bit
I/O devices; if this becomes a huge problem we can allocate in ZONE_DMA32
(but watch out for compatibility with older kernels). This change was needed
for the cachesize limits to make sense; if we switch to DMA32 we need to
modify the limit tests as well.

- Remove bogus check for DEV_SHUTDOWN in the dtr; the flag won't be set
if we're coming from a ctr error path