r203: - More stylistic cleanups
- Have sha1_compress() clear the workspace before returning.  This doesn't
seem to measurably affect speed, which sort of makes sense since the
workspace is already be in the L1 cache.  We'd like to be able to only clear
it on the last iteration before the caller returns, but since the workspace
is stored on the stack, there's no good way do that. So we do it
unconditionally to avoid leaking information about the input data.

- Add asmlinkage to sha1_compress() prototype