r256: - Convert lev1 to use a large sparse file rather than directories of files. Track chunk lengths in lev1 since the filesystem will no longer do it for us.
**** Local cache must be cleaned before updating to this version ****

- Make the fid cache not actually track fids.  It's just an LRU tracker
now.  (It's going away soon, so I'm not going to bother to rename it.)

- Distinguish between dirty, modified, and modified-this-session in chunk
state.  Previously, any chunk which had been written this session would
be written back any time the chunk was closed, even if it was clean.
Modified-this-session is only used for end-of-session statistics, while
modified is used to set the dirty bit in the cache header.

- Eliminate CHUNK_STATUS_RW flag, since we no longer need to distinguish
between read-only fds and read-write fds.

- Have strip_compression() mark newly-stripped chunks modified but
not modified-this-session, so that the chunk is properly uploaded but
doesn't factor into session statistics

- Enable large file support for vulpes.  The small-file access code
still needs to be audited for 32-bit offsets, but it's not urgent.
Be sure to "make clean" before building if you have old object files
in your working copy.

- Add stub function to copy chunks into directories-of-files for upload --
this isn't filled in yet, so uploads don't work.

- Expand cache header to a full 512-byte block, since we'll probably need
the extra reserved storage