r308: - Have convergent no longer request keys for chunks which are going to be overwritten in a whole-chunk write.  (It wasn't doing the disk I/O, but it would still request the keys.) There's no need to do so, and it makes it harder for userspace to tell what's going on.
If a whole-chunk write experiences a write failure, there are now two cases:

* Write fails to a chunk which vulpes doesn't think is dirty.  The chunk
won't be picked up during checkin, but the write failure will have been
returned to the block device user.

* Write fails to a chunk which vulpes *does* think is dirty.  The write
failure will have been returned to the block device user, and vulpes will
notice the hash mismatch at checkin.

Both of these are acceptable so long as vulpes is checking tags at checkin.
What we really need is reporting of I/O errors from the kernel back to
vulpes, but that'll have to wait.  (And it might not prevent the need for
the checks, since consistency would still break in case of a crash.)

- Have vulpes mark a chunk present if it receives a key update from the
kernel (to support the above)

- Allow userspace to reply to a key request with an I/O error (in case of
network failure, tag check failure, etc).  This error will be cached, so
this is really only for hard errors.

- Have vulpes produce I/O errors on retrieve_chunk() failure

- Bump chardev protocol version