r337: - Eliminate chunk-dirty flag from cache file header.  Unfortunately, keyring comparisons are the only certain way to determine which chunks have changed in the presence of keyring or cache corruption.  Keeping this flag around just adds extra state that must be kept consistent, and increases the number of inconsistency issues we must be able to deal with.  So, since the infrastructure now exists in vulpes for keyring comparisons, we now just stick to this approach when doing uploads and checks.
- Eliminate dirty-chunk count from cache file header

- Move printing of local cache statistics message from perl client (which
no longer has sufficient information, with the removal of the dirty-chunk
count) into the new vulpes "examine" mode.  Make the old "check" mode
functionality an optional part of "examine" so the client doesn't need to
call vulpes twice in this case.

- Have vulpes keep track of the number of valid chunks at runtime, rather
than retrieving and generating this value only for cache header reads and
writes.  When initially reading the cache header, we don't trust the stored
valid-chunk count, but generate it by counting valid bits.

- Split mark_cdp_modified() into session and non-session variants so
that the initial cache-header read doesn't factor into the shutdown
statistics

- Add get_prev_keyring() helper function

- Free prev-keyring on cache shutdown

- Fix warning in client cleantmp()