r388: - Add TFM_NR_SUITES and TFM_NR_COMPRESS defines
- Add helper functions to produce the bit index of a suite or compression
algorithm from a bitmask with a single bit set

- Replace increasingly-unwieldy suite_lookup()/compression_lookup() with an
array of structs containing suite and compression metadata, indexed by the
bit number of the suite or algorithm.  Provide suite_info() and
compress_info() helper functions to return the proper struct given a bitmask
with a single bit set.

- Include constant data (cipher blocksize and hash length) in the info
arrays which were previously obtained from CryptoAPI.  The hash length
should never change for a given algorithm.  The blocksize might, but must be
a part of the definition of a suite, so this is still safe.

- Remove transform metadata fields from nexus_dev, and get the information
from suite_info()/compress_info() instead.

- Don't set suite/default_compression/supported_compression in nexus_dev
until the values provided in the register ioctl have been validated;
this prevents invalid memory access by the sysfs attribute handlers in
a brief window while the dev ctr is running

- Add missing validation check for compresstype == 0