r394: Convert NEXUS_CRYPTO_* and NEXUS_COMPRESS_* bitmask defines into enums, since they turn out to make more sense as indexes than as bitmaps.  This includes the following changes:
- Add a new cryptomask_t for the supported-compression bitmap, which
is set/tested with something like (1 << NEXUS_COMPRESS_LZF)

- Use __u8's for the crypto and compress values in nexus.h, since the
compiler is allowed to pick the size of an enum and we need to know the size
in advance.  Use enum types for everything in the module proper.

- Add sanity checks to nexus.h to ensure build errors if there are
field-size problems

- Remove suite-or-algorithm-is-supported runtime sanity checks.  Every suite
or alg defined in the enum must be supported by Nexus.

- Throw a build error if the suite/compress description tables don't contain
exactly one entry per enumerated constant.  See previous point.

- Remove obsolete suite_index() and compress_index()

- Update userspace test program.  Add distinction between NEXUS_COMPRESS_*
constants and on-disk format, since the enumerated constants have changed
and shouldn't be used in persistent storage anyway.  Define the on-disk
values to be equal to the old bitmask constants to preserve backward
compatibility.  Print out the NEXUS_* values rather than the on-disk values
in the verbose logs.