r397: - Move transform buffers from the dev structure to a new struct nexus_tfm_state, which is embedded in the dev structure for now. Arrange nexus_tfm_state to be able to support multiple simultaneous crypto suites.
- Split transform_alloc() and transform_free() into functions which
allocate and free individual suites and compression algorithms; the original
functions are now just wrappers which handle per-device tasks such as
parameter validation

- Allocate and free compression bounce buffers based on refcount of
algorithms which are using them

- Add MAX_CHUNKSIZE constant setting the size of the vmalloc'd bounce
buffers used for compression/decompression.  This is a hack: we always
allocate this amount of memory even if we could allocate less.