r403: - Allocate per-thread transform buffers in thread.c.  The device ctr registers interest in a particular set of transforms, and thread.c keeps track of the "interest sets" of each transform, which are bitmaps of the devices which are interested in the transform.  When a device registers interest in an unallocated transform, it is allocated.  When an allocated transform no longer has any interested devices, it is freed. If an allocation fails at registration time, the registration is backed out and the device ctr fails.
- On a CPU hotplug event, allocate or free buffers for every transform
which has an interested device.  If the allocation fails, its thread is not
started, so each successfully started thread is guaranteed to be able to
service any transform request.  (Note that we already refuse to allow a CPU
to be unplugged if it is the only remaining CPU running a Nexus thread.)

- Update comment with information on CPU_LOCK_ACQUIRE/RELEASE