r507: - Allow client to take more than one fcntl lock.  Maintain a stack of locks and release them all (in reverse order) at shutdown.
- Use a separate fcntl lock for hoarding, to allow hoarding to occur in
parallel with other activities.  Take both the main and hoard locks
when doing "isr clean -a", since deleting the hoard cache out from
under "isr hoard" can cause it to report success when the hoard cache
is still missing chunks.

- Add a "description" parameter to acquire_lock() so that the error
message produced on a lock conflict specifies which lock could not be
acquired