r524: Command line argument changes:
- No longer accept --userid as a synonym for --user

- The parcel version is now specified with -r/--rollback instead of
-v/--version, since the convention is to use -v for "verbose" and --version
to print the version of the command line tool.  The use of -r is compatible
with user expectations from revision control systems, where it refers to
"revision" rather than "version".  In ISR we call them "versions"; thus the
back-formation "--rollback".  Note that "rollback" doesn't make much sense
for "isr checkp", but checkp is going away with the CAS version of the
server.

- Use -v for verbose instead of -V