r588: Compile fixes for 2.6.20:
- Update supported-version check

- Add our own typedef to replace the deprecated kmem_cache_t.  We can't
just use "struct kmem_cache" because it was "struct kmem_cache_s" before
2.6.15.

- Update workqueue calls to deal with the elimination of the private data
member.  Add compatibility wrapper for INIT_WORK and typedef for the
parameter to the work function.

- Unconditionally include <linux/freezer.h> from kcompat.h for >= 2.6.20.
We could have done the version test in thread.c, which is the only source
file that needs it, but this is easier to read.  And freezer.h isn't very
big.