r647: - Add cipher suite using sha1 tag generation but no encryption.  Don't make crypto calls or calculate keys when this suite is in use.
Note: This is sort of hacky; it should be stable but isn't optimized. The
real solution would be to allow different hash algorithms for tag and key,
and special-case the null crypto and hash algorithms to avoid the
scatterwalk overhead.  This would also change the userspace interface to
Nexus, since the hash length for tag and key might be different.

- Change log messages in __chunk_tfm() not to assume that we're doing
encryption

- Minor comment cleanup in __chunk_tfm()

- Add a parameter to setup mode for nexus/tools/test which selects the
cipher suite to be used.  This causes an incompatible change to the store
file format.

- Hardcode the path to /dev/openisrctl in nexus/tools/test rather than
expecting it as a command line parameter in setup mode

- Have nexus/tools/test bail out when it detects an old-format store file