r693: - Unblock SIGINT after checkout is complete.  This prevents SIGINT from having inconsistent handling during run(), depending on whether we did a checkout beforehand.
- Have isr_sget() use system() rather than mysystem(), and check the
curl return value to see whether it died on a signal.  If so, isr_sget()
returns the new error code $Isr::EINTR.  Update callers to propagate/handle
this code.

- Allow users to hit ctrl-C during download of data files during checkout.
Call clean() on the (inconsistent) parcel directory if this occurs.  This is
a one-off hack; we need better error-handling infrastructure overall.

- Change the logic in checkout() to not have two nearly identical copies of
the isr_sget() call