Fix race with simultaneous checkout of multiple parcels
- Have checkout() use $parcel-parcel.cfg and $parcel-nonce rather than
parcel.cfg and nonce when storing files temporarily in the $isrdir root

- Have hoard() use $parcel-parcel.cfg.hoard rather than
$parcel-parcel.cfg.$$ when creating a temporary parcel.cfg in $isrdir.
Adding the pid is not necessary, since we make sure that multiple hoards
to the same parcel cannot run simultaneously.  (We do need to use
"...cfg.hoard" rather than "...cfg" because hoard and checkout have no
locks in common.)  Removing the pid improves consistency with checkout()
and ensures that stale parcel configs don't accumulate in $isrdir in
error cases.