nexus: Don't free request queue lock on device shutdown
Don't make the request queue lock part of struct nexus_dev, since the
request queue may outlive the nexus_dev if the queue itself has outstanding
references, and some of those reference holders may want to use the queue
lock.

- Kernel versions >= 2.6.12 have a mechanism for including the request queue
lock in the request queue structure directly, so we use that trick if
available.

- For kernel versions < 2.6.12 we use a single static lock for all Nexus
devices.  In kernels < 2.6.10 this approach will be slightly racy with
regard to module unloading -- according to Jens Axboe, the block layer only
holds onto the request queue for short periods.  In 2.6.10 and above,
though, it would be very racy, since the CFQ scheduler can hold a queue
reference for arbitrarily long periods before trying to use the queue lock.
So, in 2.6.10 and 2.6.11, we prevent the Nexus module from being unloaded at
all.