Nexus: rename "unwedge" => "action"; add force-unregister command
- Add "zap" unwedge command which forcibly shuts down the Nexus device. This
will cause further I/O on the blockdev and chardev to error out, and poll()
on the chardev will return POLLERR.  This can be used to recover the system
when the userspace process is stuck in D-state.  Vulpes will probably get
confused, but at least it'll be killable.

- Rename "unwedge" sysfs attribute to "action" in anticipation of future
expansion.  Rename {cache,threads,elevator} commands to
kick-{cache,threads,elevator}.

- Have the test program print a message and exit if the Nexus device
is unregistered out from under it