Nexus: Fix kernel-user deadlock at vm_dirty_ratio
The Linux kernel has a fairly simplistic formula for controlling the number
of dirty pages in the system.  There are two variables,
dirty_background_ratio and vm_dirty_ratio, which default to 10%/40% in older
kernels and 5%/10% in >= 2.6.22.  When the number of dirty pages in the
system exceeds the background ratio, background page writeout is started
through pdflush.  When the number of dirty pages exceeds the dirty ratio,
processes which attempt to dirty a page begin to do writeout of existing
dirty pages themselves, and will block in D-state until the dirty page count
drops below the dirty ratio.  For kernels >= 2.6.19, this includes processes
which attempt to dirty a clean page in a shared writable mmap area.

There is an additional complication: if the dirty ratio is greater than half
the amount of memory *not* consumed by mmapped pages, the ratio is lowered
to half the amount of unmapped memory (but never less than 5%).  Thus, if
there is a lot of mmapped memory in the system, processes are likely to hit
the dirty ratio much sooner.

This mechanism causes problems when the system hits the dirty ratio, many of
the dirty pages belong to a Nexus block device, and the userspace key
manager tries to dirty a page.  In this case the userspace process will
block on page writebacks which can never complete because Nexus will be
unable to load and store their chunk keys.  This problem is triggered by
heavy disk I/O from a guest parcel, coupled with large parcel memory
allocations (which are mmapped, and so decrease the dirty ratio).  (The test
case I used was an OS upgrade in a 512 MB guest, running on a 2 GB host with
an additional 512 MB guest running.  This setup reliably produced deadlocks
within a few minutes.)

The proper solution is to allow Nexus to perform writebacks even if the
userspace keyring manager is temporarily unresponsive.  The workaround
employed here, instead, is to set the PF_LESS_THROTTLE process flag for the
keyring manager process; this gives it an additional 25% headroom in the
dirty ratio before it blocks on writebacks.  As a result, the keyring
manager will be able to proceed after other dirtying processes have been
throttled.  If the keyring manager dirties enough pages to exceed the
additional headroom, however, deadlock will still occur.

Note: In the case where multiple processes or threads share the same chardev
fd, or where the fd is passed between processes, it is impossible to
properly set and clear PF_LESS_THROTTLE on every process/thread that needs
it.  An ioctl has been provided to set the flag on processes/threads other
than the one which first registered the device.  (Note that the flag is
inherited across fork() and clone().)  Individual threads in a process may
not have the flag properly cleared when the chardev is unregistered.  Also,
unregistering the device will always cause the flag to be cleared, even if
another device is registered to the same process.

Patches exist against the upstream kernel which make the dirty-page
throttling more intelligent: the amount of dirty data allowed to exist
against a given backing device varies with the device's speed, and processes
dirtying pages against a given device are not forced to wait for a different
device's writeout.  The latter change should prevent this particular
problem, since the keyring manager will never write to the Nexus block
device.