Nexus: update for 2.6.22
- Remove scale argument from bioset_create_wrapper(), since bioset_create()
in 2.6.22 no longer scales bio_vec mempool sizes.  For older kernel
versions, pass a scale argument large enough to prevent any scaling from
occurring.

The reasoning for removing "scale" in 2.6.22 is that biosets only have to be
large enough to allow progress, which is typically only a request or two.
For Nexus to make progress, though, it needs to complete whole chunks, which
may be composed of more than one request.  Multiple Nexus devices may also
perform interleaved bio allocations, so that we need to allocate more than
the minimal number of bios in order to write out a complete chunk.  Given
all of this, and the fact that we don't really *have* to be able to make
progress under OOM -- the OOM killer will kick in, if nothing else -- it's
not clear how to size our bioset.  So... we stay with the numbers we've
already been using, which hopefully have a little more headroom than
necessary.