libvdisk: Build both 32-bit and 64-bit versions on 64-bit machines
- Have configure check whether we're on an x86_64 system.  If so, and if the
compiler supports the -m32 switch, build and install both 32-bit and 64-bit
versions of libvdisk, since vmware-vmx is 32-bit but the other VMware
programs are 64-bit.  If we're on x86_64 but -m32 isn't supported, print a
warning and only build the 64-bit libvdisk.

- Add "compile" script to toplevel .gitignore, since some versions of
Automake create this after the changes described above