Add script to upgrade a parcel to 0.9 format
Add beginnings of 0.8 => 0.9 upgrade script.  This is intended to be
run on the server, and does not upgrade in place: all data is copied from
the old parcel directory to a new directory which is created by the script.
(Almost all of the parcel data needs to be modified during the upgrade
anyway, and the original parcel can be kept as a backup.)  As of now, the
script does the following things:

- Create the destination directory and copy over the lockholder log
and the "last" symlink

- Read the keyroot from parcel.cfg and write it to a temporary file

- Decrypt the config tarball for each version and, if necessary, the hex
keyring

- Upgrade the keyring for each version to SQLite and remove the old-format
keyrings (from the destination directory)