locksrv/getconfig: Use get_parcelcfg_path(); remove -u option
- Use get_parcelcfg_path() rather than generating the path by hand

- Remove strange mandatory -u argument which is required to be equal to the
basename of the user's home directory.  If the username can be obtained from
the environment, then passing it on the command line isn't necessary.  (It's
not useful as a sanity check, since the client is connecting to a specific
username on the server anyway.) There's no reason to allow even
administrators to run this script with alternate usernames, since all it
does is cat a file.