pk: Expect parcel.cfg in directory given by --parcel option
- Add --parcel command line option giving the directory containing the
parcel.cfg file.  It is unreasonable to expect the client to copy (or move)
the parcel.cfg into the cache directory, because the cache directory may not
exist in all cases where the parcel.cfg would.  We also don't want to assume
that the parcel directory is the parent of the cache directory, because the
layout of ~/.isr, above the individual cache and last directories, is the
province of the Perl client and not PK.  We continue to place the
parcelkeeper.* state files in the cache directory, because they're specific
to a running cache and not to the parcel as a whole.  (For example, there
could be multiple running instances for the same parcel if snapshotting is
implemented; again, this is the province of the client code.)

- Update client to pass the --parcel option when starting PK