client: Eliminate HTTPSSH.pm demand-loading
- Always source HTTPSSH.pm at startup, and never do it from
parse_parcelcfg().  We only have one protocol, and even if we didn't, the UI
wrapper is the wrong place for protocol-specific code.  All of the protocol
support will eventually move into Parcelkeeper, whose job it is to hide
these details from the client.  Plus, with the current arrangement, we're
constantly downloading and parsing parcel.cfg files just to get access to
the functions we need to call.

- Remove gratuitous parcel.cfg fetches and parses from listparcels(),
checkparcel(), and clean()