Separate finding a cached chunk from reclaiming the LRU chunk.
In most cases we just want to find already cached chunkdata. The
only place where we want to reclaim an unused/idle slot is in
reserve_chunks. But in that case it is better to first pin down
anything that is already cached before we start evicting. Any
eviction may drop a cached chunk that we also we need for this IO.