pk: Store uuid in ASCII rather than binary format
- Store state.uuid in string format rather than binary.  We don't use the
UUID directly in very many places, so it's not a space issue as it is with
tags and keys; on the other hand, the string form is easier to work with. We
still run it through libossp-uuid in order to validate it.

- Make parcels.uuid in the hoard cache index a string rather than a blob,
for the same reasons

- Make state.uuid a pointer to malloc'd memory, rather than an array

- Remove include of uuid.h from defs.h, and add includes of system headers
in other places to compensate