pk: Grab a reference every time the parcel hits a hoard chunk
- Always add a parcel reference to a chunk whenever the parcel expresses
interest in the chunk via hoard_get_chunk() or hoard_put_chunk().  This
approach better ensures that chunks relevant to parcel A, but which have
been fetched by parcel B *after* the reference list was most recently
synchronized from A's keyring, are not garbage-collected if B loses interest
in the chunk.  This is not perfect: if A never hits a chunk that is relevant
to A but has been hoarded by B, then the chunk could be discarded before
anyone notices that A is also interested.

- Log the tag if we fail to get a chunk reference