mkrevision.sh: Refresh git index before checking for dirtiness
- Refresh the git index before checking whether to add the -dirty flag
to the revision string.  This is necessary on Coda because the Coda client
which creates/updates a file sees different stat information for that file
than other clients do, so that a tree that looks clean on one client may
appear dirty on another.  So, we have git do a little more work to determine
whether any files have *actually* changed before deciding the tree is dirty.