client: Change local cache directory structure
- Remove last-to-cache copy of memory image at resume.  This operation adds
considerable I/O overhead on every resume immediately following a checkout
or a discard, in order to make "isr discard", an infrequent operation, fast.
Instead, we decrypt the memory image directly into cache and use it from
there, ignoring "last" entirely.  On "isr discard", we re-decrypt the memory
image from the encrypted original saved in the hoard cache.

- Move the contents of the "cache" and "last" directories into their parents
and eliminate the directories.  The only purpose of "last" was to provide a
pristine keyring and memory image for keyring comparisons and discard
operations.  Discard operations can be satisfied from encrypted data in the
hoard cache (with regard to the memory image), while a pristine copy of the
keyring can be kept in the parcel directory under a different name.

- Add "clean cache" flag file to indicate that no checkin is necessary
because the parcel has not been resumed.  This was previously determined by
checking for the existence of the "cache" directory, which was created at
first resume.

- Update PK callouts to match the corresponding changes in PK, made in
an earlier commit

- Don't refuse to garbage-collect the hoard cache if the parcel has not
been resumed since checkout