locksrv: Add isr_runserv wrapper script
- Add wrapper script which calls the server script specified by its first
argument.  This is the first in a series of changes which will make
isr_runserv the sole entry point into the server code, place isr_runserv in
the PATH and the rest of the server code in $sharedir/openisr, and thus
eliminate the need for the client to know the path at which the server code
was installed (without filling the PATH with a bunch of individual ISR
server commands).