vulpes: Add shared memory segment for dirtometer
Add POSIX shared memory segment containing a map of the state of each chunk
in the virtual disk.  This map is created only in run mode and contains two
bits for each chunk, as follows:

00: Chunk not present in local cache, not accessed this session
01: Chunk present in local cache, not accessed this session
10: Chunk present and accessed this session
11: Chunk present and modified this session

The memory segment simply contains an array of these two-bit quantities,
with no headers or other data.  Updates occur as soon as Vulpes receives new
information from Nexus.  Writes to the shared memory segment by other
processes will be ignored, except that the information in the segment will
no longer be consistent.  If the segment cannot be created for some reason,
Vulpes will continue without it.