nexus: Fix rare low-memory deadlock in nexus_setup_io()
- Fix rare low-memory deadlock, most likely seen on pre-2.6.18 systems with
only one crypto thread.  (That is, systems with SMP and HT disabled.)

It turns out that mempool_alloc() never blocks in the underlying allocator;
if the initial allocation attempt fails and the pool is empty,
mempool_alloc() itself blocks waiting for an existing allocation to be freed
back into the pool -- even if more system memory has become available in the
meantime.

We use mempool_alloc() in two places: directly in nexus_setup_io() and
indirectly via bio_alloc_bioset().  The latter is safe: once a bio is
allocated by a crypto thread, it is unconditionally shipped off to be
executed, and after completion, the bio is freed in the context of the endio
callback.  Thus an allocated bio will eventually be freed, even if the
allocating thread is blocked.  The nexus_setup_io() usage, however, is not
safe: if we block there, and there's only one crypto thread, we'll stay
blocked there forever, since we can't free an io without executing in the
crypto thread.

The fix is to call mempool_alloc() in nexus_setup_io() with __GFP_WAIT
masked out.  We already have our own backoff-and-retry mechanism for
allocation failures in nexus_setup_io(), so we just use that rather than
trying to make mempool_alloc() handle retries for us.

Starting with 2.6.18, mempool_alloc() contains a hack which will retry the
allocation after blocking for five seconds.  So the bug isn't fatal on newer
kernels, just very annoying.  That hack will probably be removed in the
future, however.