client: Name parcel directories after the parcel's UUID
- Convert the .isr directory layout from .isr/$userid/$parcel to .isr/$uuid.
This should prevent conflicts between parcels on different servers with
the same userid and parcel name.

- Remove $isrhome (== ~/.isr), and convert $isrdir from ~/.isr/$userid to
~/.isr

- Allow the UUID to be specified on the command line instead of the parcel
name.  This will work with the existing isr subcommands whenever the parcel
is already checked out.

- Rename parceltmp-$userid-$parcel-$$ directory to parceltmp-$$, since
$userid-$parcel is both unnecessary and non-unique

- Name the lockfiles after the UUID instead of the parcel name.  Add a
"checkout lock" for use during the checkout phase, when we don't yet know
the UUID.  This approach preserves correctness at the cost of not allowing
simultaneous checkout of two identically-named parcels.

- Don't take a local lock in checkparcel mode, since it's server-side-only

- Name the session log after the parcel UUID.  If the UUID isn't available,
write to ~/.isr/isr.log as an interim measure.