openisr-config: Remind user that the initscript must be run at boot
- After a successful run, remind the user that the initscript must be
configured to run at boot.  This is necessary because the client asks the
user to run openisr-config if Nexus isn't loaded.  The common case is that
the modules haven't been built for the user's kernel, but if the user
installed from source, it could be that they never configured their system
to load the modules at boot.  In this case, openisr-config will fix the
problem (by loading the modules) but it will recur after reboot.  For
simplicity, we avoid clever solutions and just ask the user to check the
initscript.  Note that this message will only be displayed if they're using
*this* version of openisr-config, which is only installed when installing
from source.