client: Add support for -pre strings in Nexus version check
- Add support for suffix strings of the form "-pre", "-rc2", etc., in the
OpenISR version number.  The comparison of Nexus and client release strings
is done such that 0.8.3 < 0.8.4-arbitrarysuffix < 0.8.4.  We don't parse the
suffix (-rc1 is equal to -rc7) but when the final release comes around the
user will be prompted to upgrade if they're using a prerelease kernel
module.  This seems like a good compromise.

- Rework the check_for_nexus() version-checking logic to use a cleaner
method