client: Don't allow disconnected mode if only local cache is populated
- For disconnected mode, require that the hoard cache is fully populated,
rather than requiring that either the hoard cache or the local cache is
fully populated.  If the local cache is fully populated, then under most
conditions the hoard cache will be also, since we fill in the hoard cache as
we populate the local cache.  (The exception is if the hoard cache was
explicitly flushed using "isr rmhoard".)  On the other hand, if the local
cache is full but the hoard cache is not, an "isr discard" (which is
supposed to be safe in disconnected mode) will cause the parcel to no
longer be runnable disconnected.

- Remove hdksize(), since it is now unused