client: Don't redownload parcel metadata in disconnect()
- When calling hoard() via disconnect(), it seems silly to download and
decode the same metadata twice in a row, into two different directories.
So, we take the parcel lock during disconnect() and add a parameter to tell
hoard() that it's safe to use the existing parcel directory rather than
creating its own.  (Since the parcel lock is taken, we know that clean() or
checkin() in another instance won't delete the parceldir out from under
hoard().)