pk: Do low-level hoard cache ops on their own DB connection
- Make low-level hoard cache operations (i.e., not the hoard modes) operate
on a separate database connection.  This fixes the failure of
copy_for_upload() to insert modified chunks into the hoard cache:
copy_for_upload() was calling hoard_put_chunk() in the middle of a SELECT,
hoard_put_chunk() wants to begin and commit two different transactions,
and SQLite doesn't support nested transactions.  We can't make transactions
the responsibility of hoard_put_chunk()'s caller, since it drops and
re-acquires a transaction in the middle, and we don't really want to have
copy_for_upload() read one row at a time or build an in-memory data
structure to cache the SELECT results.  So the best solution seems to be
to make hoard.c use a separate database connection entirely; this way,
callers of the hoard cache functions can have their own database activity
going on, and as long as they don't have an open transaction which has
touched any hoard cache tables, the hoard cache code won't care.